Conditional Roaming Encounter framework, Legendary Birds, and Mew. implements the core framework for the Conditional Roaming Encounter system. currently, the system fires once every 7000 world ticks, for every player, and then has a 20% chance to actually select an encounter to try, if it succeeeds, and can spawn one roamer successfully per day. the aforementioned numbers can be tampered with to affect the frequency of and how many Roamers can appear. they can be found in `globals.js`. currently, the Roamer registry list is randomly indexed, meaning there is equal weighting to it, and any Roamer can attempt to spawn at any time and place, there are no preconditions for attempted spawns. if pokemon of the Roamer's species is already in the player's party, that Roamer will not spawn, and instead it will either run a custom function defined by the `alreadyOwns` key in roamer data, or if that does not exist, have a 50% chance to start over with a different roamer. Roamers can be set as flyers by defining a `flyingHeight` value in their data, they will attempt to spawn the given amount of blocks off the ground, and iterate downward for a surface to spawn on. currently no roamers can actually fly on their own, or atleast don't do it often. they are also currently spawned at the absolute default of `/pokespawnat`, so no levels or other property changes yet, but i will fix that by letting you run a custom function to set any properties.
includes an implementation of the Legendary Birds and Mew
Legendary Birds have an `alreadyOwns` function that will reroll into another member of the trio, so having two of the three will reroll spawns for the two you have into spawns for the third, making it 3x as likely to appear.
currently the birds don't seem to meet their condition, or atleast Moltres doesn't, probably got something to do with the biometag check code i stole from the kube discord
Mew does not have the biome condition right now because i forgot to implement it lmao, but since that seems to be broken anyway it's fine as it is.

3 custom Commands are also included.
`/forceroaminglegendaryspawn` will force a Roamer to spawn near the player (given or executor), ignoring its conditions, but not ignoring the `alreadyOwns` functions or reroll.
`/testencountercondition` will test if the player (given or executor) meets the Condition for a Conditional Encounter.
`/resetdailyroamercounter` will reset the `dailyRoamerSuccesses` tracker value of the player. (given or executor)
the former two have a `species` field which must be filled with the key name for the species data in the Conditional Encounter maps, which mostly use the species name, full lowercase, no mod prefix.
the `testencountercondition` command requires an `encounterType` field which should be `roamer` or (technically anything right now but) `static`, used to determine what encounter map it pulls from.